pyproject.toml
[project]
name = "autowork"
version = "0.1.0"
description = "autowork is a autogen sample"
authors = [
    {name = "zdzwsz", email = "zdzwsz@126.com"},
]
dependencies = [
    "pyautogen>=0.2.16"
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
chat_date = "python src/autowork/tests/main.py chat_date"
chat_file = "python src/autowork/tests/main.py chat_file"
chat_weather = "python src/autowork/tests/main.py chat_weather"
chat_rag = "python src/autowork/tests/main.py chat_rag"